Factorize common code in Tiles
  - Creation of a common BaseTile managing side panel, search and selection
  - Updating DocumentTile to use BaseTile
  - Creation of MediaTile which is the base from Audio and Video tiles
  - Suppression of stylesheets not needed anymore
