--- libs/search-widget/src/tiles/_text-tile.scss
@@ -1,96 +1,0 @@
-.sw-tile {
-  .thumbnail-container {
-    position: relative;
-  }
-
-  &.expanded {
-    padding-left: var(--rhythm-2);
-    padding-right: var(--rhythm-2);
-    transition: padding-right var(--transition-moderate);
-
-    .resource-viewer-container {
-      height: calc(100vh - var(--header-height) * 2);
-      overflow-x: hidden;
-      overflow-y: auto;
-      position: relative;
-    }
-
-    .side-panel {
-      display: none;
-    }
-  }
-}
-
-@media (min-width: 648px) {
-  .sw-tile.expanded {
-    --side-panel-button-size: var(--rhythm-4);
-    --side-panel-width: var(--rhythm-36);
-    --side-panel-gap: var(--rhythm-1);
-    --tile-padding-right: var(--rhythm-2);
-
-    padding-right: var(--side-panel-button-size);
-
-    .resource-viewer-container {
-      height: calc(100vh - var(--header-height) - var(--rhythm-1));
-    }
-    .search-result-paragraphs {
-      --search-results-margin-bottom: var(--rhythm-1_5);
-
-      background-color: var(--color-light-stronger);
-      margin-bottom: var(--search-results-margin-bottom);
-      max-height: calc(
-        100vh - var(--header-height) - var(--find-input-height) - var(--find-input-margin) -
-          var(--search-results-margin-bottom)
-      );
-      overflow: auto;
-    }
-    .side-panel {
-      display: flex;
-      gap: var(--side-panel-gap);
-      position: absolute;
-      right: calc(var(--side-panel-width) * -1 + var(--side-panel-button-size) + var(--tile-padding-right));
-      top: var(--header-height);
-      transition: right var(--transition-moderate);
-
-      .side-panel-button {
-        align-items: center;
-        background-color: var(--color-dark-stronger);
-        color: var(--color-light-stronger);
-        display: flex;
-        height: var(--side-panel-button-size);
-        justify-content: center;
-        width: var(--side-panel-button-size);
-      }
-
-      .side-panel-content {
-        width: calc(
-          var(--side-panel-width) - var(--side-panel-button-size) - var(--side-panel-gap) - var(--tile-padding-right)
-        );
-      }
-      .find-bar-container {
-        line-height: var(--side-panel-button-size);
-      }
-    }
-
-    &.side-panel-expanded {
-      padding-right: var(--side-panel-width);
-
-      .side-panel {
-        right: var(--tile-padding-right);
-      }
-    }
-  }
-}
-
-@media (min-width: 1025px) {
-  .sw-tile.expanded {
-    --side-panel-button-size: var(--rhythm-5);
-    --side-panel-width: var(--rhythm-46);
-  }
-}
-
-@media (hover: hover) {
-  .sw-tile.expanded .side-panel-button {
-    cursor: pointer;
-  }
-}
