--- libs/search-widget/src/tiles/audio-tile/AudioTile.scss
@@ -1,2 +1,0 @@
-@import '../tile';
-@import '../media-tile';
