--- libs/search-widget/src/tiles/base-tile/DocumentTile.scss
@@ -1,2 +1,0 @@
-@import '../tile';
-@import '../text-tile';
