--- libs/search-widget/src/tiles/video-tile/VideoTile.scss
@@ -1,2 +1,0 @@
-@import '../tile';
-@import '../media-tile';
